Properly fix the logo issue
Apparently, the problem happened when the header was rendered around the
verbatim environment – something in the way verbatim works breaks
\input.

It is not exactly clear what the problem is, but a good workaround is to
read the file when we definitely know that it is safe.

So, we do not \input the file each time.
Instead read it into a variable once, and then just use this variable.

Tip from: https://tug.org/pipermail/texhax/2021-September/025327.html